refactor: Relocating dto file
