+++ src/main/java/com/example/spark/domain/wma/api/PredictionController.java
@@ -1,7 +1,7 @@
 package com.example.spark.domain.wma.api;
 
 
-import com.example.spark.domain.flask.dto.YouTubeDataCache;
+import com.example.spark.domain.youtube.dto.YouTubeDataCache;
 import com.example.spark.domain.wma.service.PredictionService;
 import com.example.spark.domain.youtube.dto.YouTubeAnalysisResultDto;
 import com.example.spark.global.response.SuccessResponse;
